Move sponsor room objects
